pyproject.toml
[project]
name = "newsdim"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "polars>=1.40.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "sentence-transformers>=5.4.1",
    "tqdm>=4.67.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/newsdim"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.5",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
